refactor: split workspace watch actions into focused modules
