+++ apps/desktop/src/store/workspace/watch/actions.test.ts
@@ -1,10 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest"
-import { createWorkspaceActionTestContext } from "./workspace-action-test-helpers"
-import {
-	collectRefreshDirectoryPaths,
-	createWorkspaceWatchActions,
-	replaceDirectoryChildren,
-} from "./workspace-watch-actions"
+import { createWorkspaceActionTestContext } from "../actions/workspace-action-test-helpers"
+import { createWorkspaceWatchActions } from "./index"
 
 const { invokeMock, getCurrentWindowMock, listenMock } = vi.hoisted(() => ({
 	invokeMock: vi.fn(),
@@ -20,7 +16,7 @@ vi.mock("@tauri-apps/api/window", () => ({
 	getCurrentWindow: getCurrentWindowMock,
 }))
 
-describe("workspace-watch-actions", () => {
+describe("watch/actions", () => {
 	beforeEach(() => {
 		invokeMock.mockReset()
 		getCurrentWindowMock.mockReset()
@@ -47,92 +43,6 @@ describe("workspace-watch-actions", () => {
 		expect(unwatch).toHaveBeenCalledTimes(1)
 		expect(originJournal.clearWorkspace).toHaveBeenCalledWith("/ws")
 		expect(getState().unwatchFn).toBeNull()
-	})
-
-	it("collectRefreshDirectoryPaths collapses to top-most parent directories", () => {
-		const paths = collectRefreshDirectoryPaths("/ws", [
-			"docs/a.md",
-			"docs/sub/b.md",
-			"archive/c.md",
-		])
-
-		expect(paths).toEqual(["/ws/docs", "/ws/archive"])
-	})
-
-	it("collectRefreshDirectoryPaths treats sibling prefixes as distinct directories", () => {
-		const paths = collectRefreshDirectoryPaths("/ws", [
-			"a/file.md",
-			"a/sub/child.md",
-			"a-archive/other.md",
-		])
-
-		expect(paths).toEqual(["/ws/a", "/ws/a-archive"])
-	})
-
-	it("replaceDirectoryChildren updates only target directory subtree", () => {
-		const entries = [
-			{
-				path: "/ws/docs",
-				name: "docs",
-				isDirectory: true,
-				children: [
-					{
-						path: "/ws/docs/old.md",
-						name: "old.md",
-						isDirectory: false,
-					},
-				],
-			},
-			{
-				path: "/ws/keep.md",
-				name: "keep.md",
-				isDirectory: false,
-			},
-		]
-
-		const nextChildren = [
-			{
-				path: "/ws/docs/new.md",
-				name: "new.md",
-				isDirectory: false,
-			},
-		]
-
-		const updated = replaceDirectoryChildren(
-			entries,
-			"/ws",
-			"/ws/docs",
-			nextChildren,
-		)
-
-		expect(updated[0]?.children).toEqual(nextChildren)
-		expect(updated[1]).toEqual(entries[1])
-	})
-
-	it("replaceDirectoryChildren swaps root entries when workspace root is targeted", () => {
-		const entries = [
-			{
-				path: "/ws/old.md",
-				name: "old.md",
-				isDirectory: false,
-			},
-		]
-		const nextRootEntries = [
-			{
-				path: "/ws/new.md",
-				name: "new.md",
-				isDirectory: false,
-			},
-		]
-
-		const updated = replaceDirectoryChildren(
-			entries,
-			"/ws",
-			"/ws",
-			nextRootEntries,
-		)
-
-		expect(updated).toEqual(nextRootEntries)
 	})
 
 	it("watchWorkspace clears stale unwatchFn when watch start fails", async () => {
